requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
python-dotenv==1.0.0
mistralai==0.3.0
asyncio==3.4.3
redlines==0.5.2
google-genai==1.38.0

